Updates for Svc/CmdDispatcher corresponding to UT cmake udpate for protected/private
